# specs/001-moex-ticker-update/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MOEX Ticker History API
  version: 1.0.0
  description: Returns daily MOEX history and the latest executed trade available without subscription credentials. History selects the board primary on each trading date. Under normal operating conditions (10 concurrent clients issuing 10 requests per second in total), at least 95% of successful responses complete in under one second end to end.
  license:
    name: UNLICENSED
    identifier: LicenseRef-Proprietary
servers:
  - url: /
paths:
  /v1/moex/{SYMBOL}:
    get:
      operationId: getMoexTickerHistory
      summary: Get daily history for a MOEX ticker
      security: []
      parameters:
        - name: SYMBOL
          in: path
          required: true
          description: MOEX instrument security identifier.
          schema:
            type: string
            minLength: 1
            pattern: '^[A-Za-z0-9._-]+$'
      responses:
        '200':
          description: Daily records available without subscription credentials, selected from the board that was primary on each trading date and ordered ascending by date. No available records is represented by an empty array. The `facevalue` property is populated from the current MOEX LOTSIZE for the selected board, or null when unavailable; the property name is retained as requested.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DailyMarketRecord'
              example:
                - date: '2013-03-25T00:00:00Z'
                  close: 73.35
                  high: 75.05
                  low: 73.21
                  volume: 120300
                  facevalue: 1
        '400':
          description: The symbol is malformed, unsupported, or cannot be resolved to a primary board.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: MOEX ISS failed, access was denied, or returned unusable data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: The history response could not be read from or persisted to the service's history store.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/moex/{SYMBOL}/quote:
    get:
      operationId: getMoexTickerLatestTrade
      summary: Get the latest executed trade for a MOEX ticker
      description: Fetches the latest trade from MOEX ISS for this request. Returns a one-element array using the six-field history record shape. A valid empty trades block for a recognized ticker returns one record with all six fields null. Quote data is never reused from a prior request.
      security: []
      parameters:
        - name: SYMBOL
          in: path
          required: true
          description: MOEX instrument security identifier.
          schema:
            type: string
            minLength: 1
            pattern: '^[A-Za-z0-9._-]+$'
      responses:
        '200':
          description: One latest-trade record, or one all-null record when a valid ISS response contains no latest trade for the recognized ticker.
          content:
            application/json:
              schema:
                type: array
                minItems: 1
                maxItems: 1
                items:
                  $ref: '#/components/schemas/LatestTradeQuoteRecord'
              example:
                - date: '2026-10-08T08:41:57Z'
                  close: 280.34
                  high: null
                  low: null
                  volume: 1
                  facevalue: null
        '400':
          description: The symbol is malformed, unsupported, or unknown.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: MOEX ISS failed, access was denied, or returned malformed trade data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    LatestTradeQuoteRecord:
      type: object
      required: [date, close, high, low, volume, facevalue]
      properties:
        date:
          type: [string, 'null']
          format: date-time
          description: Trade execution time normalized to UTC; null when ISS returns a valid empty trade block.
          example: '2026-10-08T08:41:57Z'
        close:
          type: [number, 'null']
          description: Latest executed trade price; null when ISS returns a valid empty trade block.
        high:
          type: 'null'
          description: Always null for a latest-trade record.
        low:
          type: 'null'
          description: Always null for a latest-trade record.
        volume:
          type: [number, 'null']
          description: Latest executed trade quantity; null when ISS returns a valid empty trade block.
        facevalue:
          type: 'null'
          description: Always null for a latest-trade record.
    DailyMarketRecord:
      type: object
      required: [date, close, high, low, volume, facevalue]
      properties:
        date:
          type: string
          format: date-time
          description: Trading date at midnight UTC, preserving the MOEX trading date.
          example: '2013-03-25T00:00:00Z'
        close:
          type: [number, 'null']
          description: Closing price; null when unavailable from the source.
        high:
          type: [number, 'null']
          description: Daily high; null when unavailable from the source.
        low:
          type: [number, 'null']
          description: Daily low; null when unavailable from the source.
        volume:
          type: [number, 'null']
          description: Daily traded volume in instrument units; null when unavailable.
        facevalue:
          type: [number, 'null']
          description: Current LOTSIZE from the MOEX instrument reference for the primary board selected for this record; null when unavailable. The property name is retained as requested and does not indicate historical FACEVALUE.
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: moex_unavailable
            message:
              type: string
              example: MOEX market data is temporarily unavailable.